+++ crates/core/component/ibc/src/component/msg_handler/connection_open_ack.rs
@@ -12,8 +12,8 @@ use penumbra_chain::component::StateReadExt as _;
 use crate::{
     component::{
         client::StateReadExt as _,
-        client_counter::validate_penumbra_client_state,
         connection::{StateReadExt as _, StateWriteExt as _},
+        ics02_validation::validate_penumbra_client_state,
         proof_verification, MsgHandler,
     },
     IBC_COMMITMENT_PREFIX,
