+++ crates/core/component/ibc/src/component/msg_handler/connection_open_try.rs
@@ -17,9 +17,9 @@ use penumbra_chain::component::StateReadExt as _;
 
 use crate::component::{
     client::StateReadExt as _,
-    client_counter::validate_penumbra_client_state,
     connection::{StateReadExt as _, StateWriteExt as _},
     connection_counter::SUPPORTED_VERSIONS,
+    ics02_validation::validate_penumbra_client_state,
     MsgHandler,
 };
